+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 9. `jQuery`: Upgrade to version 3.4.1
 10. `FontAwesome`: Upgrade to 5.10.2
 11. `Card`: Adding `href` to navigate.
+12. `Table`:Fix rowspan issue when use grouping together with removeDuplicate
 
 ## Version 4.0.0